requirements.txt
pip==23.1.2
mysql-connector-python==8.1.0
streamlit==1.25.0